refactored and added new test, which checks if testcontainer is ok and running
